+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+## [1.0.2] - 2026-01-29
+
+### Summary
+
+fix(core): resolve Changelog, [Unreleased], Added and 7 more
+
+### Changes
+
+- build: update Makefile
+- config: update goal.yaml
+- docs: update EXAMPLES.md
+- docs: update QUICK_REFERENCE.md
+- docs: update README
+- update pactfix-py/pactfix/analyzer.py
+- update pactfix-py/pactfix/analyzers/__init__.py
+- update pactfix-py/pactfix/analyzers/docker_compose.py
+- update pactfix-py/pactfix/analyzers/gitlab_ci.py
+- update pactfix-py/pactfix/analyzers/kubernetes.py
+- update pactfix-py/pactfix/analyzers/terraform.py
+- config: update deployment.yml
+- update pactfix-py/tests/fixtures/terraform/main.tf
+- update pactfix-py/tests/test_analyzer.py
+- update playwright-report/index.html
+
 ## [1.0.1] - 2026-01-29
 
 - docs: update CHANGELOG.md